zstd_safe

Enum CParameter

#[non_exhaustive]
pub enum CParameter {

    TargetCBlockSize(u32),
    CompressionLevel(CompressionLevel),
    WindowLog(u32),
    HashLog(u32),
    ChainLog(u32),
    SearchLog(u32),
    MinMatch(u32),
    TargetLength(u32),
    Strategy(Strategy),
    EnableLongDistanceMatching(bool),
    LdmHashLog(u32),
    LdmMinMatch(u32),
    LdmBucketSizeLog(u32),
    LdmHashRateLog(u32),
    ContentSizeFlag(bool),
    ChecksumFlag(bool),
    DictIdFlag(bool),
    NbWorkers(u32),
    JobSize(u32),
    OverlapSizeLog(u32),

}

A compression parameter.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

TargetCBlockSize(u32)

Target CBlock size.

Tries to make compressed blocks fit in this size (not a guarantee, just a target). Useful to reduce end-to-end latency in low-bandwidth environments.

No target when the value is 0.

CompressionLevel(CompressionLevel)

Compression level to use.

Compression levels are global presets for the other compression parameters.

WindowLog(u32)

Maximum allowed back-reference distance.

The actual distance is 2 power โthis valueโ.

HashLog(u32)

ChainLog(u32)

SearchLog(u32)

MinMatch(u32)

TargetLength(u32)

Strategy(Strategy)

EnableLongDistanceMatching(bool)

LdmHashLog(u32)

LdmMinMatch(u32)

LdmBucketSizeLog(u32)

LdmHashRateLog(u32)

ContentSizeFlag(bool)

ChecksumFlag(bool)

DictIdFlag(bool)

NbWorkers(u32)

How many threads will be spawned.

With a default value of 0, compress_stream* functions block until they complete.

With any other value (including 1, a single compressing thread), these methods directly return, and the actual compression is done in the background (until a flush is requested).

Note: this will only work if the zstdmt feature is activated.

JobSize(u32)

Size in bytes of a compression job.

Does not have any effect when NbWorkers is set to 0.

The default value of 0 finds the best job size based on the compression parameters.

Note: this will only work if the zstdmt feature is activated.

OverlapSizeLog(u32)

Specifies how much overlap must be given to each worker.

Possible values:

0 (default value): automatic overlap based on compression strategy.

1: No overlap

1 < n < 9: Overlap a fraction of the window size, defined as 1/(2 ^ 9-n).

9: Full overlap (as long as the window)

9 < m: Will return an error.

Note: this will only work if the zstdmt feature is activated.