Struct zstd_safe::DCtx

pub struct DCtx<'a>(_, _);

A Decompression Context.

The lifetime references the potential dictionary used for this context.

If no dictionary was used, it will most likely be 'static.

Same as DStream.

Implementations

impl DCtx<'static>

pub fn create() -> Self

impl<'a> DCtx<'a>

pub fn decompress(&mut self, dst: &mut [u8], src: &[u8]) -> SafeResult

Wraps the ZSTD_decompressDCtx() function.

pub fn decompress_using_dict(
&mut self,
dst: &mut [u8],
src: &[u8],
dict: &[u8]
) -> SafeResult

pub fn decompress_using_ddict(
&mut self,
dst: &mut [u8],
src: &[u8],
ddict: &DDict<'_>
) -> SafeResult

Wraps the ZSTD_decompress_usingDDict() function.

pub fn init(&mut self) -> usize

Wraps the ZSTD_initCStream() function.

Initializes an existing DStream for decompression.

pub fn reset(&mut self) -> SafeResult

Wraps the ZSTD_resetDStream() function.

pub fn load_dictionary(&mut self, dict: &[u8]) -> SafeResult

pub fn ref_ddict<'b>(&mut self, ddict: &DDict<'b>) -> SafeResult where
'b: 'a,

pub fn ref_prefix<'b>(&mut self, prefix: &'b [u8]) -> SafeResult where
'b: 'a,

pub fn set_parameter(&mut self, param: DParameter) -> SafeResult

pub fn decompress_stream(
&mut self,
output: &mut OutBuffer<'_>,
input: &mut InBuffer<'_>
) -> SafeResult

Wraps the ZSTD_decompressStream() function.

pub fn in_size() -> usize

Wraps the ZSTD_DStreamInSize() function.

Returns a hint for the recommended size of the input buffer for decompression.

pub fn out_size() -> usize

Wraps the ZSTD_DStreamOutSize() function.

Returns a hint for the recommended size of the output buffer for decompression.

pub fn sizeof(&self) -> usize

Wraps the ZSTD_sizeof_DCtx() function.